Isolate commands, colors and errors
