+++ src/colors.js
@@ -1,0 +1,6 @@
+module.exports = new Object({
+    "error" : 0xff0000,
+    "sticky" : 0xffff00,
+    "success" : 0x00ff00,
+    "info" : 0x3399ff
+});
