+++ src/errors.js
@@ -1,0 +1,6 @@
+module.exports = new Object({
+    "invalid_channel" : "Example: 798815345905106945 (Right-click channel and Copy ID, or tag the channel #a-channel)",
+    "no_stickies_channel" : "There are no stickies for this channel.",
+    "no_stickies" : "There are no stickies to list.",
+    "no_sticky_id" : "There are no stickies with the ID for that channel."
+});
