+++ src/stickies.js
@@ -1,0 +1,6 @@
+const {Stickies} = require("./sticky.js");
+
+if (!global.stickies)
+    global.stickies = new Stickies();
+
+module.exports = global.stickies;
